requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
boto3==1.37.21
botocore==1.37.21
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///croot/click_1698129812380/work
cloudpickle @ file:///croot/cloudpickle_1683040006038/work
contourpy @ file:///croot/contourpy_1700583582875/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz @ file:///croot/cytoolz_1701723583781/work
dask @ file:///croot/dask-core_1715838601404/work
exceptiongroup==1.2.0
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.25.0
fsspec==2024.6.1
ftfy==6.3.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
grad-cam==1.5.2
grpcio==1.62.1
h11==0.14.0
hf-xet==1.1.3
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.32.4
idna @ file:///croot/idna_1666125576474/work
imagecodecs @ file:///croot/imagecodecs_1695064943445/work
imageio @ file:///croot/imageio_1707247282708/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib-resources @ file:///croot/importlib_resources-suite_1704281845041/work
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keopscore==2.1.1
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lazy_loader @ file:///croot/lazy_loader_1718176737906/work
lightning-utilities @ file:///home/conda/feedstock_root/build_artifacts/lightning-utilities_1721751555632/work
llvmlite==0.42.0
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath==1.3.0
munkres==1.1.4
networkx==3.2.1
numba==0.59.1
numpy==1.23.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
open_clip_torch==2.32.0
openai-clip==1.0.1
opencv-python==4.7.0
packaging==24.2
pandas==2.2.2
partd @ file:///croot/partd_1698702562572/work
pillow==11.0.0
ply==3.11
protobuf==4.21.12
ptflops==0.6.9
pybind11==2.13.6
pydantic==2.7.0
pydantic_core==2.18.1
pykeops==2.1.1
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-transformers==1.2.0
pytz==2024.1
pywavelets @ file:///croot/pywavelets_1705049820073/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2024.11.6
replicate==0.25.1
requests @ file:///croot/requests_1707355572290/work
s3transfer==0.11.4
sacremoses==0.1.1
safetensors==0.5.2
scikit-image @ file:///croot/scikit-image_1682528309703/work
scikit-learn==1.5.1
scipy==1.13.0
seaborn==0.13.2
seam-carving==1.1.0
sentencepiece==0.2.0
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1692548726750/work
termcolor==2.2.0
threadpoolctl==3.5.0
tifffile @ file:///croot/tifffile_1695107451082/work
timm==0.6.12
tokenizers==0.21.1
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
toolz @ file:///croot/toolz_1667464077321/work
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics @ file:///home/conda/feedstock_root/build_artifacts/torchmetrics_1715837061745/work
torchvision==0.21.0+cu118
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.67.1
transformers==4.52.4
triton==3.2.0
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.0.2
yacs==0.1.8
zipp @ file:///croot/zipp_1704206909481/work
